backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Cors==3.0.10
psycopg2-binary==2.9.6
requests==2.28.2
python-dotenv==1.0.0
email-validator==2.0.0
pytest==7.3.1